requirements.txt
pupa==0.9.0
https://github.com/opencivicdata/python-opencivicdata-django/zipball/master
lxml
sh 
pytest==4.2.0
pytest-mock==1.10.1
requests-mock==1.5.2